requirements_versions.txt
accelerate==1.9.0
blendmodes==2024.1.1
clean-fid==0.1.35
diffusers==0.34.0
diskcache==5.6.3
einops==0.8.1
facexlib==0.3.0
fastapi==0.112.4
GitPython==3.1.44
gradio_imageslider==0.0.20
gradio_rangeslider==0.0.8
gradio==4.40.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
inflection==0.5.1
jsonmerge==1.9.2
kornia==0.8.1
lark==1.2.2
loadimg==0.5.0
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
open-clip-torch==3.0.0
peft==0.16.0
piexif==1.1.3
pillow-avif-plugin==1.5.1
Pillow==10.4.0
protobuf==4.25.8
psutil==7.0.0
pydantic==2.8.2
pytorch_lightning==2.5.1
resize-right==0.0.2
safetensors==0.5.3
scikit-image==0.25.2
setuptools==78.1.0  # temp fix for compatibility with some old packages
spandrel-extra-arches==0.2.0
spandrel==0.4.1
tomesd==0.1.3
torch
torchdiffeq==0.2.5
torchsde==0.2.6
tqdm==4.67.1
transformers==4.54.1